Remove mask and old_mask values from irq_info
We don't need them anymore, because unroutable irqs just don't get touched
anymore, and we have the assigned_obj pointer to gather our mask value from
during activation.

Note - This removal does necessecitate the removal of affinity hint, but we'll
be reimplementing that soon, as the prior policy was rather inflexible.