Fix segfault on non-numa enabled systems
If a system doesn't support numa, we have an emtpty top level of the balancing
tree (and we get a null dir pointer in build_numa_nodes to boot).  Fix this by
checking for that null condition and adding the unspecified numa node to the
list of all numa nodes.

See https://bugzilla.redhat.com/show_bug.cgi?id=746159 for details

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>