Fix oops on non-numa enabled systems
Some systems aren't numa enabled, but still have devices that indicate they
belong to numa node 0 (rather than -1 as they should).  Thats a kernel problem
but we still shouldn't crash as a result (which we will when device get a NULL
numa_node pointer as a result of trying to find non-existant node 0).  This
patch fixes the crash by always returning the unspecified node pointer when numa
is disabled

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>