Default the cache_domain mask to that of the cpu creating it
Currently, if a cpu creates a cache domain in do_one_cpu, and if that cpu has
not cache information ennumerated in sysfs, the result is a garbage cache_domain
mask, because the cache attribute set doesn't exist, leading to a cache mask of
whatever garbage is on the stack.  Default the behavior to give new
cache_domains the same mask that the child creating the cpu has.  This will
create a situation in which each cpu belongs to one unique cache domain in the
absence of other affinity information, which seems like a sane default

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>