Add a SIGINT signal handler
This allows for a clean teardown by setting the loop condition to false
when the signal is received. We then free the memory as we would in
oneshot mode, making it much easier to use valgrind on a long-running
process and not be overwhelmed by false positives.

With this patch applied, the following reports no memory leaks after
letting it run for five minutes:

G_SLICE=always-malloc valgrind --suppressions=valgrind.supp
    --leak-check=full --show-possibly-lost=yes --show-reachable=yes
    ./irqbalance --debug

Signed-off-by: Dan McGee <dan@archlinux.org>
Signed-off-by: Neil Horman <nhorman@tuxdriver.com>