numa: only ignore the unspecified node during placement when numa is available
Serj Kalichev reported a problem with irqbalance on non-numa systems, indicating
that irqs were not balanced because the placement algorithm ignores the
unspecified node.  We need to not do that, but only in cases where we don't have
any other nodes.  If we have a valid numa system, then we want to always choose
one of the real nodes over the unspecified one, but in non-numa systems we can
and should allow the selection of the unspecified (and only) node

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
Reported-by: Serj Kalichev <serj.kalichev@gmail.com>