apply affinity hint also if the current policy is subset
--hintpolicy=subset chooses an object that has a non-empty intersection
with the affinity hint, but it never restricts the object's CPU mask
with the hint itself.  As a result, there is no guarantee that the
object's CPU mask is a subset of the hint.

This is visible for interrupts whose balancing policy is not BALANCE_CORE.
For example, if there is only one cache domain and the interrupt's policy
is BALANCE_CACHE, the chosen object will correspond to "all CPUs" and
the affinity hint will be effectively ignored.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Neil Horman <nhorman@tuxdriver.com>