Improve rescan ability for newly allocated interrupts
When an interrupt is registered after irqbalance starts we just add it as a misc
interrupt, even if its something that we want to balance.  Instead, lets add the
irq to a temp list, force a rescan, then make sure the newly discovered irqs get
added properly on the rescan.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>