numa: Reset unassigned node on tree rebuild
http://code.google.com/p/irqbalance/issues/detail?id=38

This issue noted that on a tree rebuild, the unassigned node has pointers to the
interrupt list and children list that is left over from the previous tree state.
We erroneously follow these lists even though they are no longer allocated,
leading to memory errors

Fix it my re-initializing to the startup state.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>