correct standard deviation for 2 cpu case
When running on a system with just two cpus, its guaranteed that both cpus load
will be within a bessel corrected standard deviation of the average, meaning no
balancing will ever occur.  In these cases, cut the s.d. in half so that we are
more sensitive to load imbalances.  Also change the find_overloaded_objs macro
to a function, as its getting too unwieldy

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>