Fix infinite rescan loop when new non-PCI IRQs appear
As reported in bug:
http://code.google.com/p/irqbalance/issues/detail?id=58

If a new non-PCI based IRQ arrives between scans (one not found in
sysfs) then we would add the new IRQ to a new_irqs_list and set
need_rescan=1.  The rescan would clear the interrupts_db, add all sysfs
based IRQs, and finally add _just_ the newly found IRQ to the
interrupts_db, but not any of the original non-PCI IRQs.  This means on
the next parse_proc_interrupts() scan we would find the original non-PCI
IRQs, add them to the new_irqs_list, set need_rescan=1 and the cycle
would repeat between the newly found IRQs and the originals.

The new_irq_list concept was added in fd24d8f30012a "Improve rescan
ability for newly allocated interrupts" with what appears to be the
intention of ensuring we balance any newly found IRQs.  This change
reverts most of that commit but attempts to keep the intention intact by
calling force_rebalance_irq() for each non-PCI based IRQ found.  It also
still sets need_rescan=1 if this is not the first pass through to
hopefully catch and correctly classify any sysfs based IRQs.

Signed-off-by: Shawn Bohrer <sbohrer@rgmadvisors.com>
Signed-off-by: Neil Horman <nhorman@tuxdriver.com>