consider banned cpus earlier during placement
Dave Sullivan and Petr Holasek reported that irqbalance didn't set affinity
properly when IRQBALANCE_BANNED_IRQS was set to a single cpu.  The problem
turned out to be that numa node assignment for an irq is done without
consideration for the banned cpu set.  The result is that irqs can be assigned
to a node that has no elligible cpus, and as a result, those irqs don't have
their affinity set.  Fix this by restricting node assignment to nodes that have
an elligible cpu as defined by the unbanned_cpus bitmask

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>