Adjust default hintpolicy value
affinity_hint values are something of a holdover from prior to when irqbalance
was re-written.  Previous to the rewrite irqbalance could not determine with
great accuracy the device to which an msi(x) irq was associatied, the node it
was local to, etc, and so was not able to balance it well.  kernel affinity_hint
values were created to work around that by just telling user space where to put
an interrupt.  However, since the rewrite, irqbalance is perfectly capable of
parsing all information about an irq out of sysfs, and can make superior policy
decisions about balancing according to user input, over kernel suggestions.  As
such, allow users to honor affinity_hint, but by default ignore it

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>