Set default deepestcache to 2
Been meaning to do this for a while.  By default having a deepestcache value of
ULONG_MAX causes irqbalance to always find the deepest cache level in a systems
which causes it on some systems to think that all cpus share a cache, and as
such, that no balancing is needed.  Rectify this such that the default deepest
cache level is defaulted to 2

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>