classify.c: Initialize struct user_irq_policy
commit 2139f8a8952c9eeea6e77e5587f13cd17724e23d "make hintpolicy user
policy configurable" introduced a regression where if the user does not
provide a policy script the user_irq_policy struct will not be
initialized.

The common pattern for get_irq_user_policy() is that a struct
user_irq_policy is allocated on the stack, get_irq_user_policy() is
called, and then pol.ban == 1 is tested.  Without a policy script this
became undefined since the struct was not being initialized.

Signed-off-by: Shawn Bohrer <shawn.bohrer@gmail.com>