Fix inconsequential bug in bitmap_scnprintf
Fix incorrect code in bitmap_scnprintf that could cause it to give truncated
output when nmaskbits > 32 and nmaskbits % 32 != 0.  In practice this problem
was never encountered because NR_CPUS % 32 == 0.  So this change fixes the
code for the sake of correctness even though the bug did not actually cause a
real problem.