Support PCI quirks in classify code
Current irqbalance uses PCI class code to determine the irq class
for each of devices. This logic could work well on most of devices.

However, the problem is, there are some PCI devices don't follow
the PCI spec. That caused the improper irq class code assigned to
these devices. Usually, it is hard to ask HW/FW vendors to fix this
kind of bug. Some of those are not bugs, because some new PCI class
code were not defined before the new devices shipment. At same time,
it is very difficult to force a device firmware upgrade even if the
new firmware fix is available for users.

As a correct irq class is crucial for interrupt balance efficiency
and performance, we need some software workarounds for these devices.
This is also what Linux kernel PCI code does in practice.

Signed-off-by: Yong Yang <yangoliver@gmail.com>