irqbalance: fix find_best_object() for the same d->load value
find_best_object() has a problem that three or more topo_obj, which have
the same ->load value, are involved. The comparison of the ->interrupts
length are always held with the one, which is registered in best->best.
Think about the following case:

  topo_obj A, the length of ->interrupts is 3.
           B,                               1.
           C,                               2.

  and find_best_object() eats them in the order of A, B, and C.

Here, find_best_object() wrongly chooses C, instead of B, because
the length of -> interrupts of B is not compared with that of C.

To fix this, remember only the best one and compare it with the next one.

Signed-off-by: Seiichi Ikarashi <s.ikarashi@jp.fujitsu.com>
Signed-off-by: Neil Horman <nhorman@tuxdriver.com>