irqbalance: Support arm irq parsing
Add support to try to guess at platform based interrupts.  currently platform
based irqs don't provide irq->device mapping in sysfs, nor do they consistently
expose information regarding device type (which is normally used to guesstimate
interrupt volume for balancing purposes).  To overcome this, we implement some
regex matching logic to match on primarily the name in /proc/interrupts, but
optionally also in the platform device directory in sysfs.

This could be done for all arches but because:
1) /proc/interrupts is formatted slightly differently in each arch, leading to
odd behavior

2 This is being done primarily for aarch64

I'm limiting the inclusion of this code only to that arch.  The intent is to use
this until the platform device sysfs directory can be augmented to map device to
irqs in a simmilar manner to pci, and to establish some sort of class mapping,
so as to make this operation more consistent and device independent.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com