Fix string allocation issue in sock_handle
The ui interface in irqbalance uses very poor string handling to manage
its return data.  It had fixed strings of sizes 1024 and 2048 bytes to
handle any sized system, and even on a moderately sized system, we would
overrun those arrays, which were allocated on the stack, leading to
corruption.

Fix the problem by using malloc/realloc to grow the sizes of the arrays
as needed.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>