policyscript: Add ability to specify a directory for multiple scripts
The policyscript directive allows for the specifcaion of a single script
to define policy for all hardware on a system, which is good as a
site/host specific utility, but it makes for difficult work in the event
that vendors wish to provide guidance for only their own hardware (i.e.
if vendor A wants certain hardware to follow affinity_hinting without
affecting other hardware).  To manage this, lets enhance policyscript to
allow the specification of an entire directory, to which multiple
scripts can be added.  Semantics for this new directory feature are the
same as for the single script case, except that the script exit codes
have additional meaning:

exit code 0 - the script indicates that the referenced irq relates to a
device that this script recognizes and further script processing should
stop

exit code 1 - the script indicates that the referenced irq does not
relate to a device the script recognizes, and script processing should
continue

exit code >2 - the script indicates an error has occured, and any output
from it should be ignored, script processing should continue

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>