fix the problem of banmod that memory is freed before using
Currently strdupa() is used to allocate memory for irq_info's name in
collect_full_irq_list(), we know that it allocate memory from stack,
when the invoking function return, the memory will be freed. so if the
irq_info's name is invalid, it will lead to check_for_module_ban() no
correct.

  check_for_irq_ban
    check_for_module_ban(res->name) // res->name is not valid

Use strdup() instead of strdupa(), and free the memory of irq_info's
name before freeing the irq_info.

Signed-off-by: Yunfeng Ye <yeyunfeng@huawei.com>