Add strlen checking for IRQBALANCE_BANNED_CPUS
When the environment variable IRQBALANCE_BANNED_CPUS is not specifily
the cpu, we expect do not place IRQs on CPUs the kernel keeps isolated
or nohz_full.

But currently, it will go to call cpumask_parse_user and ignore the
isolated or nohz_full. so add strlen checking before calling
cpumask_parse_user.