Adjust how we determine if a cpu is online
https://github.com/Irqbalance/irqbalance/issues/159

recently brought to our attention that online cpu status isn't functional on all
arches.  Specifically on parisc, the availability of
/sys/devices/system/cpu/cpu<N>/online is in question.  The implication here is
that its not feasible to accurately determine cpu count, and as a result,
irqbalance doesn't work on that arch

Fix it by changing our online detection strategy.  The file
/sys/devices/system/cpu/online is a cpulist format file that seems to be present
accross all arches and configs.  As such, we can use this file to determine
online status per cpu reliably.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>