drop NoNewPrivs from irqbalance service
A recent update to libcapng is issuing an error in the system log,
caused by the fact that irqbalance attempts to drop capabilities when
the systemd service unit has already done so for us.  Since irqbalance
drops the caps correctly, theres really no need for us to do so via
systemd as well.  So lets drop NoNewCaps in the service unit.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>