irqbalance: Check validity of numa_node
If Linux maps an IRQ associated with an empty node (hence not exposed
under /sys/devices/system/node/), irq_numa_node(info) will be set to
NULL, leading to a core dump in place_irq_in_node().

Add a warning for this unusual condition and set ->numa_node to the
the default entry.

Signed-off-by: Cédric Le Goater <clg@kaod.org>