Implement Netlink helper functions to subscribe thermal events
Establish Netlink socket connection in a nonblocking mode and callback
notification.

Tolerate a few times on failures from receiving events. In practice, such
delivery failure is rare. But it may indicate more fundamental issues if
it repeats. So disconnect it unless the failure is transient.

Event-specific handler will be implemented in the next patch.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>