irqbalance: properly check if irq is banned
When checking if a new irq should be banned, make sure
to compare the irq's name against the list of banned modules regardless
of whether or not the new irq is in a known list of interrupts.

Therefore, for the new irq in question, pass the entire irq_info data
structure to check_for_irq_ban() instead of just the irq number.

As a result, tracking the GList of interrupts is not needed in
check_for_irq_ban() so cleanup these unused parameters across
applicable functions in the the call trace.

Signed-off-by: Nick Child <nick.child@ibm.com>