thermal: Fix the warning message
The commit febe697ac321 ("change the log level in thermal.c from error to
warning") happens to insert an unneeded message: "thermald may not be
running."

This is not true because the events come from the kernel and Netlink has
the multicast subscription model. So it has nothing to do with thermald.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>