activate_mapping: fflush the buffered data to smp_affinity
Previously irqbalance uses the return value of fprintf() to decide whether
the modification of smp_affinity is successful or not. However it is not
reliable because fprintf() is stream buffered, a fflush() should be used to
check the buffered data been successfully written into the file before the
judgement.

This patch fixes the issue by introducing fflush() after fprintf().

Signed-off-by: Tao Liu <ltao@redhat.com>