activate_mapping: do not ban irq on out of space error
If smp_affinity cannot be enforced on a given IRQ because the kernel
returns an ENOSPC error, do not ban it for the rest of the process'
life.

Only log the warning and let irqbalance try again later.

If the kernel returns any other error (like EIO when trying to change
affinity for PPI interrupts on aarch64), add it to ban list.

Fixes: 55c5c321c73e ("arm64: Add irq aff change check For aarch64 ...")
Signed-off-by: Robin Jarry <rjarry@redhat.com>