Revert "arm64: Add irq aff change check"
This reverts commit 55c5c321c73e4c9b54e041ba8c7d542598685bae.

The banned_irqs list is meant to be configured by the user, it is not
a dynamic list. Also, remove_one_irq_from_db() was freeing an irq_info
struct which was still referenced in banned_irqs and potentially in
rebalance_irq_list.

Iterating on any of these lists will cause a use after free error.

Fixes: https://github.com/Irqbalance/irqbalance/issues/267
Signed-off-by: Robin Jarry <rjarry@redhat.com>