meson: drop redundant install_man() options
It is unnecessary to set the install directory for manpages, since meson
infers this from the file extension and installs *.1 manpages to mandir
in the man1 section for you.

It is also unnecessary to tag them as "en" because that is the
default... it is likewise unnecessary to tag them with manual file
permissions since the default 644 permission is also the overridden one.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>