Version option should return 0 rather than 1
Previously invoke irqbalance with --version options, the return value
is 1 instead of 0:

$ irqbalance --version
irqbalance version 1.9.4
$ echo $?
1

It is unexpected because irqbalance have successfully returned the
version string with no errors, so 0 should be returned instead of 1.
This will confuse some automation tests.

This patch will make irqbalance return the correct value for version
option.

Signed-off-by: Tao Liu <ltao@redhat.com>