Advance the signal handler add
When there are too many devices in the /sys/bus/pci/devices directory, it
will increase the service startup time. In my environment, the
build_dev_irqs() function takes more than 30 seconds. During this time,
the service will not be able to respond to some signals (such as the -15
signal sent by systemd).

Signed-off-by: langfei <601376534@qq.com>