thermal: fix segfault due to mismatching genl attribute
Recent kernel commit 1773572863c4 ("thermal: netlink: Add the commands
and the events for the thresholds") merged in v6.13 added the
"THERMAL_GENL_ATTR_TZ_PREV_TEMP" attribute in the middle of enum
thermal_genl_attr and so all the following attrs are now offset by one
position.

This causes irqbalance to segfault during thermal event handling as it
encounters a null nlattr in the parsed attrs array in the position where
it expects to find THERMAL_GENL_ATTR_CAPACITY (which is now returned in
the next array position). The null is unchecked and passed to nla_len()
which dereferences it to obtain the nla_len, and triggers the segfault.

Fix this by removing the hardcoded attrs and instead including the uapi
header file. Also, check the parsed attr for null to prevent crashing
in case of similar kernel updates.

Signed-off-by: Anthony Iliopoulos <ailiop@suse.com>